wifi: ath12k: Use frequency instead of channel number
The current code uses a macro that handles only a maximum channel
of 11 in the 2.4 GHz band. This becomes problematic when channel 13
is configured, as failure logs are observed due to the channel limit.
To improve this, the condition is modified to use frequency in place of
the channel number. This change is beneficial as it can handle overlaps
between bands, thus preventing the observed failures.

Signed-off-by: Hari Naraayana Desikan Kannan <quic_hnaraaya@quicinc.com>
Change-Id: Id4f2d43f8ed65802b8b949e0b10b3984dee44186