QSDK: mac80211: Fix warn on if mesh metric is zero
During performence measurement, stats is disabled through boost
performence script. So, host is not receiving the metric values
from firmware and the value becomes zero during mpath finding.
This issue addressed through static mpath. But,in RvR far range
test, connection becomes unstable and getting this warnon
continuously. So, add a check to prevent this warnon if the
stats is disabled.

Patch-dependency:
	913-001-wifi-mac80211-avoid-tx-pending-queue-in-native-wifi-.patch

Change-Id: Ie132c9e506f48078087ee071a8aa608b8002cb47
Signed-off-by: Thiraviyam Mariyappan <quic_tmariyap@quicinc.com>