QSDK: wifi: ath12k:  Handle vendor command for interface offload type
Add support to handle interface offload type configuration
through vendor command from userspace tools. Use the offload
type configured to allocate respective PPE VP and map it with the
netdev. This enables the PPE HW to identify the flows and handle
the frames in the respective offload datapath configured.
Make the DS enabled by default

Patch-depdendency: 951-c-QSDK-wifi-ath12k-support-DS-for-SLO-and-MLO-AP-modes.patch

Change-Id: Iab2ab1940d2bb4b203c3d80a5f68f8459adc0497
Signed-off-by: Balamurugan Mahalingam <quic_bmahalin@quicinc.com>