wifi: ath12k: Fix control app interface creation
For non single wiphy architecture, interface combination for multi hardware
is published in wiphy information but this parameter should be available only
for the single phy architecture.

In control_app, we fetch the hw_idx to confirm the non single wiphy architecture
Due to misleading information, interfaces are not created and testcases
are failing.

Add check to advertise the interface combination for multi hardware only if
MLO module parameter enabled.

Fixes: 029ea3cf ("wifi: ath12k/cfg80211: Enabling the SLO in non-single
		  wiphy architecture")

Depends on: 882-wifi-ath12k-Enabling-the-SLO-in-non-single-wiphy-archi.patch

Change-Id: I82aaae65cf72e615c5001f08a566a49b3936366a
Signed-off-by: Karthik M <quic_karm@quicinc.com>