Fix race condition in tx completion handler during vif teardown
This commit addresses a race condition that occurs when traffic (DL) is
running in the background and hostapd/wpa_supplicant are terminated.
The vif, accessed in the tx completion handler for debug stats, is deleted
as part of the bring down sequence. If any dp tx pending buffer is ongoing
and the vif is not set to NULL, unexpected behavior can occur while handling
the pending tx buffers.

The issue where the tx_desc_used_list was not being removed during the
mac_vif_unref operation is resolved. The vif is now unreferenced during the
down operation for the interface for which the tx completions are not yet
handled.

This is achieved by obtaining the used tx_desc from the allocated pages and
unreferencing the skb_cb->vif for which the down operation is requested.

Change-Id: Ic0f191d50159fb3b0828d249cfdd17c9dd217496
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>