wifi: mac80211: Add the MLD vif support to handle beacon loss
In ieee80211_beacon_loss function triggers the QoS null
packet for some time from STA to check whether the AP is
alive. But there is no support for Multi Link VAP's on
to send the Null packet. So remove the multi link condition
and allow the STA to send the Qos Null packets even the vif
is MLD.

Dependency: [None]
Patch-work: [None]

Fixes: Ic102d5a6ce6 ("ath12k: Add support for STA mode to
trigger the disconnection by using ieee80211_beacon_loss.")

Change-Id: I1c9519e53b4d91a2bb0de80d1bfe9d8fae025a55
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>