ath12k: fix firmware crash due to invalid peer nss value
Currently, if the access point receives an association
request containing an Extended HE Capabilities Information
Element with an invalid MCS-NSS, it triggers a firmware
crash

This issue arises when EHT-PHY capabilities shows support
for a bandwidth and MCS-NSS set for that particular
bandwidth is filled by zeros and due to this, driver obtains
peer_nss as 0 and sending this value to firmware causes
crash

Address this issue by implementing a validation step for
the peer_nss value before passing it to the firmware. If
the value is greater than zero, proceed with forwarding
it to the firmware. However, if the value is invalid,
reject the association request to prevent potential
firmware crashes.

Patch-work: NA
Patch-dependency: NA
Feature-code: 320MHz

Change-Id: I7c73181e376f00dd3303839fddf0128ddd9dded3
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>