wifi: ath12k: Fix MCS for 20 MHz only
When STA sends the assoc request with capped MCS
capabilties till 9 or 11, AP should communciate
with the capped MCS values 9 or 11. During 20MHZ
2GHz operation, AP is responding with MCS values
of 11 or 13 which is higher than the STA's capability.

AP calculates the MCS values for 80MHz Bandwidth rather
than 20MHz which is sending wrong values.

Fix by sending correct MCS capabilities to FW when AP is
configured in 20 MHz bandwidth in 2 GHz radio.

Depends-on: 572-ath12k-prepare-EHT-peer-assoc-parameters.patch

Change-Id: I33388a82f326b346f7f4f113f30265c3ca742e8a
Signed-off-by: Karthik M <quic_karm@quicinc.com>
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>