wifi: ath12k: Fix processing 4addr NULL frame in WMI rx path
Currently link_valid is set for all hw mode while processing
4addrress null frame in WMI rx path.
But link_valid shall be set only for EHT mode and it will be
dropped in mac80211 if the field set for other modes.

Below changes are done to correctly process 4addr null frame,
so that frame will not be dropped by mac80211.

* link_valid is set only for EHT mode
* link_id is set for EHT mode as peer->link_id
* using ieee80211_rx_napi() to pass pubsta to avoid sta links
  mismatch in MLO case
* add check to drop data frames other than 4address null data
  frame, to avoid warn_on_once() trigger in mac80211 as link_id
  and valid_links are not set.

Fixes:  I32dbaa3f2 (ath12k: Add support to process WDS null
frame with 4addr)

Change-Id: Icc2fbf8cc6545464ea8cdf47f00cdfe9ac85d0ce
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>