QSDK: wifi: ath12k: Fix the KW issue for ppeds_irq_name out-of-bound access
Currently, in the ppeds_irq_name two-dimensional array, we are
copying more than the array size in the row index.

This will cause an array out-of-bounds issue.

To fix this, use the row index along with ppeds_irq_name
(for example: ppeds_irq_name[PPEDS_IRQ_PPE2TCL]) in the sizeof
operator.

Patch-dependency: 763-b-wifi-ath12k-add-framework-to-support-PPE-DS.patch
                  840-a-ath12k-PPEDS-DS-bring-up-for-QCN6432.patch

Fixes: I00e38f68a (wifi: ath12k: add framework to support PPE DS)
Fixes: I50bd0c085 (wifi: ath12k: PPEDS - Enable DS for QCN6432 on kernel v6.1)

Change-Id: I5991fa7b4a9cf833050bbd42572c8a7042a2cb1e
Signed-off-by: Vinith Kumar R <quic_vinithku@quicinc.com>