mac80211: re-enable multiple virtual interfaces (VIFs) support fix
The below commit has already submitted a method to fix this issue but left a
workaround when apply with some buggy scripts, which is in fact not enable in current scripts.

Commit 78e406e (mac80211: fix multiple virtual interfaces (VIFs) support, Sun Dec 3 18:11:38 2017 +0200)

Since the current mac80211 script can work well with this method,
remove the workaround and enable this fix to solve the issue.

CRs-Fixed: 3834565
Change-Id: I4e92c77ec1110258c26edd8d940e63475e6c26ac
Signed-off-by: Long Cheng <quic_longchen@quicinc.com>