wifi: ath12k: KW: Klock work fixes for ath12k.
Changes in qmi.c:

In case of failure in request_firmware the fw_entry will be NULL and
accessing the contents of fw_entry in these scenarios will lead to
crash. So moving the debug print after the check on ret value.

In case of memory-region entry is not found in DTS entry, control might
return an uninitialized value so initializing ret to -EINVAL before
returning.

Changes in debug.c:

ath12k_gstrings_stats is a 2-D array of char's with 2 strings of
size 32 each. So the index positions 32..63 are also valid. But
here dereferencing with *ath12k_gstrings_stats instead of passing
base address ath12k gstrings_stats because of which only first
element is being considered and it's length is 32.

So, changing *ath12k_gstrings_stats to ath12k_gstrings_stats so that
the whole array will be considered.

Changes in coredump.c:

The dump variable was never initialized and in the current flow
depending on some checks driver may or may not allocate memory for
dump and post this driver will check if dump is pointing to NULL
and return.

But because the variable was never initialized there is a chance that
it might be pointing to some invalid location and driver will end up
accessing it even if the memory is not allocated. Avoid this by
initializing dump to NULL.

Patch-depdendency: 881-wifi-ath12k-add-ethtool-support-for-rx-and-tx-stats
                   596-a-ath12k-add-WCN7850-QDSS-QMI-message.patch
                   596-b-ath12k-Add-support-for-QDSS-dump-collection.patch

Change-Id: Iff4a9bde77f6d03ec5c5c9722a4a4d31454b8a5f
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>