wifi: ath11k: KW: klock work fixes for ath11k
Changes in coredump.c:

If there is a failure in memory allocation, we should return without
trying to collect the dump as in case of mem allocation failure as
seg_info and segment will be pointing to NULL and deferencing them
will lead to kernel crash.

The dump variable was never initialized and in the current flow
depending on some checks driver may or may not allocate memory for
dump and post this driver will check if dump is pointing to NULL
and return.

But because the variable was never initialized there is a chance that
it might be pointing to some invalid location and driver will end up
accessing it even if the memory is not allocated. Avoid this by
initializing dump to NULL.

Changes on qmi.c:

m3_mem->vaddr is being checked for NULL and yest being passed as
argument to memcpy in skip_m3_alloc.

Moving the NULL check to the begining to avoid this issue.

Patch-depdendency: 194-ath11k-use-dev_coredumpm-API-to-collect-rddm.patch

Change-Id: I62c4eea960a77028e03c7d8dc54a14a29eaaecc6
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>