wifi: ath12k: Fill AFC PSD and EIRP in SET TPC for SP AP
Add an API ath12k_reg_get_afc_eirp_power_for_bw, to fetch the
EIRP from the channel object of the AFC payload with an input
frequency and bandwidth.

In the WMI_VDEV_SET_TPC_POWER_CMD, if the AP power mode is Standard power
and if AFC power event is received, fill the EIRP and PSD power
with the values received in the AFC response data.

Patch dependency: 333-0020-mac80211-add-support-for-6G-channels-and-regulatory.patch
Tag: 6G_REGULATORY

Change-Id: I142cd7904b0b8caabe153059ba26aa9914017bd9
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>