KW: mac80211: Fix uninitialized memory in channel switch IE parsing
Currently, while parsing the channel switch information
element, the EHT operation parameters are being calculated
incorrectly.

This is because of, memory for storing the parsed EHT
operations IE details is allocated from the heap using
the kmalloc() API. This function returns a pointer to an
uninitialized memory area, which may contain garbage
values.

Resolve this issue by replacing the use of kmalloc() with
kzalloc(), as kzalloc() will allocate memory and
automatically initialize it to zero.

Patch-Work: None
Patch-Dependency: 828-wifi-mac80211-Fix-to-update-6GHz-power-type-for-station.patch
Fixes: If77bb9550c4 ("wifi: mac80211: Fix to update HE and
EHT operation optional fields")

Change-Id: I6c569744c432b309d647c199dea7485caaf24184
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>