KW: mac80211: fix null pointer exception in ieee80211_reconfig
Currently, when the interface type is OCB, the
ieee80211_reconfig() function passes a 'link' pointer to
the ieee80211_link_info_change_notify() function. If this
'link' pointer is NULL, it results in a system crash.

This is because ieee80211_reconfig() not checks 'link'
pointer for NULL before passing it to
ieee80211_link_info_change_notify()

Fix this issue by adding a null check for 'link' pointer
before passing it to ieee80211_link_info_change_notify()
function

Patch-Work: None
Patch-Dependency: None

Change-Id: I7307b01b889f567ca3b8cff3f025579cd2229106
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>