coredump.sh : Fix incorrect dump path for QCN6432.
Currently the devcoredump framework is searching for the dump
in directory d100000.remoteproc:remoteproc_pd4:remoteproc_pd1
for both the QCN6432 radios which is wrong.Because of this it
is stuck during mini dump collection for 5 minutes.

Fix it by pointing to correct dump location.

d100000.remoteproc:remoteproc_pd4:remoteproc_pd2 ---> for external radio1

d100000.remoteproc:remoteproc_pd4:remoteproc_pd3 ---> for external radio2

Change-Id: I9781e7353d927279a287c5888d04677d7c40cb30
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>