QSDK: ath12k: add support to enable ppe vp per vif
Add support to enable ppe vp per vif.

Add support to pass the skbs that are part of accelerated flows to ppe vp path.

The rest of the rx packets take the standard linux receive path
using netif_receive_skb().

Patch-dependency: 955-QSDK-wifi-ath12k-Handle-vendor-command-for-interface-offl.patch

The subscription of 2 new fields(16 bytes) in msdu_end tlv in the rx descriptor
increases the size of the hal_rx_desc and the increase in size of hal_rx_desc
beyond 144 bytes leads to a throughput drop of 60-80Mbps in UDP UL.

The current size of hal_rx_desc is 144 bytes.
The size of hal_rx_desc with this patch is 160 bytes.

If 2 other fields(16 bytes) are unsubscribed, the size of hal_rx_desc will become
144 bytes again.

With 2 other fields unsubscribed(say sw_frame_group id, phy_ppdu_id) and
with this patch (i.e hal_rx_desc size retained as 144 bytes), no throughput
degradation is observed.

Change-Id: Iaeb94aeff2320350d1e37e66dd3e55a322af9b80
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>