wifi: ath12k: Enable umac reset for QCN6432
As the Q6 of ipq5332 is shared between all the radios of qcn6432
the umac reset interrupt line needs to be resetted for qcn6432
every time after receiving the interrupt so that the next interrupt
can be raised for that reason writing the value 0x00000100 into a pci
slot specific address.

Currently for qcn6432 MSI based interrupts are being used but for umac
reset we can't use MSI based interrupt because as per the design one
interrupt line is reserved for umac_reset interrupt so adding support
to parse it from the dts.

Change-Id: I44cebfcb0ee14b21d82e59d8d2ba1f1bb1bcfb59
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>