QSDK: wifi: ath12k: remove skb alignment logic during allocation
skb alloc returns skb->data with 64 bytes aligned address for IPQ5322 and
IPQ9574. skb->head to skb->data width is assigned based on cache size.
To align this address to 128 bytes, skb_put operation is performed for
each skb which is consuming CPU cycles. This operation is performed
inorder to align with the AXI bus size which is also 128 bytes for most
of the hardware platforms. If unaligned address is provided for AXI
transaction, AXI takes one more cycle to complete the transaction from
target to host. But CPU cycles spend for skb_put is more when compared
to that additional AXI transaction. So removing the skb_put operation in
the host to improve CPU.

Patch-dependency : None

Change-Id: I17881ec21dbaee5ec2bca951cc8e73591f4df002
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>