QSDK: wifi: ath12k: allow fast rx support for station mode
To allow fast rx in station mode, peer's vif type needs to be
validated to find whether the type is station mode or not. This
introduces additional static check in that function. Already
stats_disable, AP vif type checks which are also static checks which will
not change over runtime. So, remove this static checks from the check fast
rx and introduces the supports_fast_rx flag in the peer structure which get
set when stats disabled for AP and STA mode.

Patch-dependency: 653-002-ath12k-Allow-fast-rx-by-bypassing-stats-update.patch

Change-Id: I7a785709f89455d88a0174d694a453e90d8b4c45
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>