wifi: ath12k: Removed the EHT header reset for SLO connection
EHT MLO header is already reset and there is duplicate of MLO header reset
which trigger Firmware crash during peer association. Removed the duplicate
reset to avoid the crash during the association.

Added code to copy the EHT phy capabilities during the peer association.
This will fetch the peer capabilties and notify it to the Firmware during
association.

Change-Id: Ida61b5d266658b5ebd9a7c76b6142a8986776781
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>