QSDK: wifi: Fixes for dcs interference detect and CW
1) Enabling dcs on non AP interfaces is not supported and a the command
for the same is to be declined with a error message. But Additional
warnings and backtrace are seen.This change has been added to remove
the additional warnings and call trace seen during dcs enable on non
AP interfaces.

2) Taking mutex lock in ath12k_mac_set_cw_intf_detect() is not expected
and is leading to freezes once iw set intf_detect command is received.
Hence updated it to use lockdep_assert_held.

3) cw_info_list is used without initialising the list head leading to
crashes once CW event is received where we try to append to the
uninitialised list. Added initialisation to fix the issue.

patch-dependency: None
patch-work: None
Fixes: I6acee3318(5296735: QSDK: wifi: Add support for CW and WLAN
interference)

Change-Id: I723999b4f94b4c55a00aa39a3dbb503d94190b6b
Signed-off-by: Amit Pandit <quic_pamit@quicinc.com>
Signed-off-by: Naveen S <quic_naves@quicinc.com>