wifi: ath12k: Send paddr of QCN6432 instead of IPQ5332 for AFC
On QCN6432, when WMI_AFC_CMD is sent to target after writing the AFC
response data to the shared memory, there is a target assert seen.

When AFC memory region is allocated during bootup, the paddr is
computed as an offset of the m3dump. IPQ5332's dev node of m3 dump
is taken from the dts, instead of pebble. Sending this incorrect paddr is
causing the target assert, when AFC response is read in the shared memory.

To fix issue, use the devnode of the respective userpd of QCN6432 for
finding the paddr.

Fixes: I1b07a5deeb4 ("wifi: ath12k: Allocate shared memory for AFC payload")
Tag: 6G_REGULATORY

Change-Id: I725de3dfb3418a83a106c1a9e45f53799a2f79fe
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>