wifi: ath12k: Ensure removal of non-ML stations during recovery
In the existing code, during recovery, a low acknowledgment is reported to the
user-space for a station. This triggers a deletion operation for the station
from the user-space. However, any operation in the driver that necessitates
communication with the firmware returns -ESHUTDOWN when the
ATH12K_FLAG_CRASH_FLUSH dev_flag is set.
Remove non-ML stations when a station deletion operation is initiated
during recovery irrespective of the error code as the associated peer
object would be eliminated.

Patch-Dependency: 901-wifi-ath12k-Peer-clean-up-during-subsystem-recovery.patch
Patch-work: None

Fixes: I6a031582676 ("wifi: ath12k: Skip low ack reports for Mesh VIF")

Change-Id: I8d15b18df3d186d141a16ff3cc6edc70f30f2c62
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>