UPSTREAM: wifi: mac80211: check/clear fast rx for non-4addr sta VLAN changes
When moving a station out of a VLAN and deleting the VLAN afterwards,
the fast_rx entry still holds a pointer to the VLAN's netdev, which
can cause use-after-free bugs.

Fix this by immediately calling ieee80211_check_fast_rx
after the VLAN change.

This change is brought from upstream code for fixing the issue

https://lore.kernel.org/all/20240316074336.40442-1-nbd@nbd.name/

Signed-off-by: Felix Fietkau <nbd@nbd.name>

Change-Id: Ia7d361a84ede37dae75977a88100838c8fa4b1eb
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>