wifi: ath12k: Rectify frequency in WMI_VDEV_SET_TPC_CMD for SP AP
For a 6GHz Standard Power AP, the center frequencies of all the
subchannels of the WMI_VDEV_SET_TPC_CMD are the same as the primary
center frequency of the AP.

Since the center frequency of the current channel context is filled in
the center_freq, the center frequencies of all the subchannels are
the same.

To fix this issue, in the case of an SP AP, remove the assignment of
center_freq and let it be obtained from ath12k_mac_get_psd_channel.

Fixes: I142cd7904b0 ("wifi: ath12k: Fill AFC PSD and EIRP in SET TPC for SP AP")
Tag: 6G_REGULATORY

Change-Id: I0fdaf8a9616ad558c5c6f837e9b62dccbc2976e8
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>