wifi: ath12k: Send wdev in power mode change complete event
When an AP operates on 5 GHz and 6 GHz in SLO mode, and when
set_6ghz_power_mode command is executed on the 6 GHz interface,
in the hostapd, the AFC power mode change update complete is getting
received on the 5 GHz interface instead of the 6 GHz interface,
resulting in failure.

This is because, in the ath12k driver, for the AFC power mode change
update complete, cfg80211_vendor_event_alloc is called without a
wdev. Since there is a single wiphy architecture in ath12k, in the
hostapd, the NL event gets delivered to the first interface, in this
case it is the 5 GHz interface.

To fix this issue, send wdev in the AFC power mode change update complete
event, to indicate the power mode change complete on the 6 GHz interface.

Tag: 6G_REGULATORY
Fixes: I2dafe592977 ("wifi: ath12k: Add 6 GHz power mode switch support")

Change-Id: I5ceeeaf1b3d89e862e81f01812ac9151bb217a7b
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>