wifi: ath12k: Handle reboot when SSR recovery in progress
When SSR recovery in progress and reboot happens getting crash
as shutdown function proceed with core_deinit.
In reboot sequence, ahb_shutdown should return with a warning
that SSR recovery is interrupted due to shutdown request
when FLAG_RECOVERY is set and not proceed with core_deinit
as dependent modules might be unloaded already.

Patch-dependency: 710-001-ath12k-Add-AHB-driver-support-for-ipq5332.patch
Change-Id: Ia622b9489685f496c2e12edb3545f38c3194ec18
Signed-off-by: Aishwarya R <quic_aisr@quicinc.com>