wifi: ath12k: Fix MCS not attaining till 11 in EHT20 in 2 GHz
Currently when AP is configured in EHT20 and STA is
configured in EHT40 in 2 GHz, MCS rates are not reachable
till 11.

When STA is configured in EHT40, MCS NSS should
be set for <= 80 MHz, but MCS NSS is being set for
20 MHz only based on the peer phy mode which is EHT20_2G
in this case which in turn is decided based on AP and
STA capabilities combined.

Fix the issue by ensuring that for 20 MHz only cases the
MCS NSS should be set differently and for rest of the
cases MCS NSS should be set based on <= 80 MHz.

Patch-dependency: 572-ath12k-prepare-EHT-peer-assoc-parameters.patch

Patch-work: None

Fixes: I33388a82f326 (wifi: ath12k: Fix MCS for 20 MHz only)

Change-Id: I5a6aa2312f65f039769ce24031d810c1f3d8be8e
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>