wifi: ath12k: Rectify bandwidth in WMI_VDEV_SET_TPC_CMD for SP AP
For a 6GHz Standard Power AP, if the EIRP is reduced from 36 dBm to 26
dBm for a channel in the AFC response, the output of iw dev still shows
the tx-power as 36 dBm.

In ath12k_mac_fill_reg_tpc_info, if the AP power mode is SP,
the bandwidth (bw) passed to ath12k_reg_get_afc_eirp_power is the
current channel context’s bandwidth. However, the Channel Frequency
Index (cfi) passed to ath12k_reg_get_afc_eirp_power is the 20 MHz
sub-channel’s cfi of the AP’s bandwidth.

To fix this issue, pass bw as NL80211_CHAN_WIDTH_20 for the
20 MHz subchannels.

Fixes: I142cd7904b0 ("wifi: ath12k: Fill AFC PSD and EIRP in SET TPC for SP AP")
Tag: 6G_REGULATORY

Change-Id: I38218a2497ef6601a33564b657a1aca8d1b126d4
Signed-off-by: Hariharan Basuthkar <quic_hbasuthk@quicinc.com>