wifi: ath12k: Coalesce freq objects info received in AFC response
When an AFC response is received, the regd rules are intersected with
the frequency objects’ information received in the response. Currently,
the intersection happens in such a way that if there are two frequency
objects - 5925 MHz to 6050 MHz and 6050 MHz to 6590 MHz, and the
regd rule for SP is in the range of 5925 MHz to 6590 MHz, the regd rule
will be split into two, similar to the frequency objects.

When a command to switch the power mode to SP is issued, there is a
check to see if the current channel and bandwidth fit into any of the
SP regd rules. However, if the current bandwidth spans across multiple
regd rules, the bandwidth is incorrectly marked as not supported in SP
power mode.

To fix this issue, coalesce the frequency objects’ information received
in the AFC response so that the SP regd rules are intersected based on
the coalesced objects.

Change-Id: Id6910b8460104a3e18e31e6e8939c71ef9c50a71
Signed-off-by: Amith A <quic_amitajit@quicinc.com>