QSDK: wifi: mac80211: ath12k: enable packet counter stats in DS
Currently, packet and byte counters are not updated in iw wlanX
station dump as the stats are not updated to mac80211 from
ath12k driver due to missing config flags and sanity checks.
Fix this by modifying the sanity checks and update stats to
mac80211 in DS enabled case appropriately.

Patch-Dependency: 951-c-QSDK-wifi-ath12k-support-DS-for-SLO-and-MLO-AP-modes.patch
Patch-work: None

Fixes: I5feb78a770b ("QSDK: wifi: ath12k: Redesign DS framework")

Change-Id: I471be7f8cacc0f0c082d8d8909a3a0435441ef18
Signed-off-by: Pradeep Kumar Chitrapu <quic_pradeepc@quicinc.com>