QSDK: mac80211: Fix choosing 2G link for data traffic in MLO case
In AP_VLAN case when the skb->fast_xmit is set, it tries to
find the sta using AP's sdata using destination address which
is backhaul machine. Due to that the sta is filled with NULL
and in driver it chooses deflink for transferring the packet
which is 2G which is ipq5322 and that causes traffic stall in 2GHZ+5GHZ/6GHZ.

Fix this by find the sta using original sdata which is VLAN sdata.

Fixes: I3f6a29a9 (QSDK: mac80211: Fetch AP's sdata and vif for AP_VLAN interface)

Change-Id: I6670d78f75b4df285af203b1d41121e97238b045
Signed-off-by: Tamizh Chelvam Raja <quic_tamizhr@quicinc.com>