QSDK: wifi: ath12k: Fix unsupported bandwidth flags for 320MHz
When an AP in 5GHz is configured to operate in 240MHz using ru puncturing
the reg rules are checked for the maximum bandwidth allowed in the
frequency range in which the AP is currently operating. If the maximum
bandwidth is 160MHz, NL80211_RRF_NO_320MHZ flag is set for the current
channel and this prevents the AP from operating in 240MHz which is a
regulatory violation. Since the reg rules being received already had this
flag set for reg rules where the maximum bandwidth was 240MHz, a logic was
added to remove the flag in order to allow operation in 240MHz. But the
reg rules received no longer has this flag and hence the logic to remove
it is incorrect. This results in AP being able to operate in 240MHz even
when the maximum bandwidth is 160MHz as the flag is getting removed.

To fix this issue, remove the obsolete logic of removing
NL80211_RRF_NO_320MHZ flag in ath12k_reg_build_regd.

Fixes: Iecacb86c526 ("ath12k: update unsupported bandwidth flags")
Patch-dependency: 647-01-ath12k-Add-5GHz-240MHz-support.patch
Patch-work: None

Change-Id: Id464da537141cade29dc37c0f2ec2fa9ff04b5e0
Signed-off-by: Amith A <quic_amitajit@quicinc.com>