wifi: ath12k: Refactor ring reaping logic in tx completion
Ring reaping in tx completion is based on budget and next
entry availability. But this reaping logic cache invalidation
of entries is not done properly. So, change the reaping based
on valid entries. Addition to that, next set of new entries will
be processed if already processed entries is less than budget
until processed entries equals to budget. This will result in
better usage of CPU since new napi schedule will get avoided.

Patch dependency: 720-c-ath12k-Optimize-the-tx-completion-routine-by-avoid-a.patch

Change-Id: I94c72e92ea9e6d5a2432174bc444d765963315ff
Signed-off-by: Yuvasree Sivasankaran <quic_ysivasan@quicinc.com>