wifi: ath12k: fix correctly dma unmap paddr and free skb
Currently, if a failure occurs during DMA mapping, dma_unmap_single
is performed for the paddr of the specific skb that failed.
However, this should be done for all the queued skbs.
Additionally, if rx_desc or desc is not available, the corresponding
skb should be freed instead of freeing only the queued skb.

Hence, to address this issue unmap the DMA for
ATH12K_SKB_RXCB(skb)->paddr and queuing the skb in a list when
rx_desc or desc is not available so that we can free the failed skb
as well.

Fixes: 13ef45e0eb(955-02-ath12k-simplify-rx-buffer-ring-replenish-logic.patch)

Change-Id: Icbbd949cd8026d297e708ecd9e983488a6f5ef55
Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>