QSDK: ath12k: Update ppe_vp_num in peer struct only for PPE_VP_ACTIVE
During 4addr peer connection ppe_vp_num is being updated in
peer structure which is used in the fast rx path where the
packet give to ppe with the ppe_vp_num. Since this is updated
without checking any ppe_vp_type, this is updated for SFE
mode as well and that causing the KPI impact in rx path.
Fix this by update ppe_vp_num only if the ppe_vp_type is PPE_VP_ACTIVE

Fixes: I014d56299d7(QSDK: wifi: ath12k: Enable DS support for WDS AP_VLAN interface)

Change-Id: I1a7a93d9861d4501e07abbf0f4ec1552077b2ce9
Signed-off-by: Tamizh Chelvam Raja <quic_tamizhr@quicinc.com>