wifi: ath12k: Fix CFR capture for 80Mhz BW with 160MHz STA.
In EHT160 mode, once station is associated, the host does not the set the 80Mhz capability flag for the firmware, this causes the firmware to not capture Channel Frequency Response (CFR) for 80Mhz in EHT160 mode. Fix this issue by setting the flag for 80Mhz capability if the STA BW is greater or equal to 80Mhz in EHT160.

Fixes: 1e5aee4a23c6 (mac80211: Package upgrade to unify ath11k and ath12k driver)
Patch-dependency: None

Signed-off-by: Aditya Gawali<quic_agawali@quicinc.com>
Change-Id: I77251484f26b9c144bf9e9f8080094cfda1da2fe