wifi: ath12k: remove the arvif lookup in ath12k_wmi_bcn_tmpl
Lock is missed to add in function ath12k_wmi_bcn_tmpl while
accessing the critical data like arvif. Due to this lock miss,
system get crashed in this function. Because if reading and writing
happen at the same time, there is a chance of crash or unexpected
behavior.

Remove the arvif lookup from ath12k_wmi_bcn_tmpl as it is already
present in the previous caller of the funciton.

Patch-dependency: 691-0004-ath12k-add-critical-update-support-during-set-beacon.patch

Change-Id: I1ecde1a19e9964cfcd2873b779201cbf888c0910
Signed-off-by: Balamurugan Ramar <quic_bramar@quicinc.com>