wifi: ath12k: Avoid acquiring lock on dp rx path
In ath12k_dp_rx_deliver_msdu(), spin_lock is acquired to get arsta
for every msdu rx packet using peer. This will impact in KPI as it
will slow down the process in rx path.

Fix this by introducing a bridge peer flag in peer data structure.
Make use of this flag in the rx path, and drop the packet if any
received.

Fixes: Ie44815ff42 (wifi: ath12k: Add support for Bridge peer)

Change-Id: Idae80e0d1128fc0fda31443de3eaa9f253ce8d3e
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>