wifi: ath12k: Reuse existing vdev for scan
When initiating a scan using the iw dev scan command on an already
active interface, reuse the existing vdev created for scanning.
Avoid creating an additional vdev to prevent exhausting the maximum
number of vdevs per radio.

Fix this by checking if any vdev is already available for the
designated radio under the same MLD when triggering the scan request
from mac80211.

Patch-dependency:
       652-04-ath12k-Support-scan-for-single-wiphy-architecture.patch

Fixes: I023e5bb9b56 wifi: ath12k: Split scan request for split band device

Change-Id: Ic5022cd110ee54de50c1f3f5e133989d3234d355
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>