wifi: ath12k: Clear ag dev flag 'ATH12K_FLAG_RECOVERY'
Modify the mechanism for clearing the ag->dev_flag. Previously, the flag
was cleared when ah->num_radios matched the restart count. This lead to
failure in clearing the dev flag in case of Mode-1, where only one chip
assertion would fail to clear the ag flag.

Now, enforce the clearance of the ag dev flag `ATH12K_FLAG_RECOVERY` for
all modes, post the clearance of the dev flag for asserted ab(s).

Fixes: 14877d4201cd ("wifi: ath12k: WDS subsystem recovery fix")

Change-Id: I5ea81371b3001c65a0bc73f3f25969ff89977edd
Signed-off-by: Sidhanta Sahu <quic_sidhanta@quicinc.com>