ath9k: disable airtime fairness feature
For the CPU limitation of QCA4531, enable the airtime fairness
feature affects the Tx performance. Especiall in MCC mode,
ieee80211_schedule_txq() function is called during the queue
scheduling, and packets from different channels are mixed into the
local->active_txqs, which causing packets from the other channel
to block the transmission of the current channel.

Change-Id: I0006cbae4fd53cdd15f937ee4c226bce3aa148ea
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>