mac80211: Fix parsing of Channel Switch Announcment in 6 GHz
In the function `ieee80211_parse_ch_switch_ie`, the `new_chandef` is
constructed using parameters received in the Channel Switch Announcement
(CSA). However, the control element is currently filled only with the
width parameter, neglecting the regulatory information received in the
CSA. This omission leads to a failure in parsing the CSA Information
Element (IE), causing client disconnections.

To address this issue, update the control element to include the
regulatory information received in the CSA. This ensures accurate
parsing of the CSA IE and prevents client disconnections.

Change-Id: I2373c13eeb3508b6a153f654229ff84948aba367
Signed-off-by: Amith A <quic_amitajit@quicinc.com>